fix: fix the existence of multiple isDark variables in an app
